libpdbg/bmcfsi.c: Fix up cranking of the SBE
The self boot engine (SBE) requires clocking when first powering on
the system. The tool attempted to do this but it failed to correctly
boot a system. This seemed to be due to sending the FSI break sequence
afterwards.

Instead move the FSI break sequence to the initialisation so that this
tool can be used to boot the system.

Signed-off-by: Alistair Popple <alistair@popple.id.au>