src/main.c: Fix SPR encoding
The internal getreg/putreg functions encode the register to read by
assuming registers > R31 are special purpose registers (SPR). However
when reading/writing the SPR this offset was not removed and so the
incorrect SPR was read (or written). This is fixed by subtracting the
correct offset.

Also fixes a minor whitespace issue.

Signed-off-by: Alistair Popple <alistair@popple.id.au>