Fix thread status checks during instruction ramming
Instructions can only be rammed into a thread is all threads on a
chiplet are quiesced or sleeping. This was not properly enforced which
leads to system checkstops when ramming threads if the chip isn't
fully quiesced. This adds a check to ensure all threads are quiesced.

It also changes the startchip/stopchip behaviour to stop all threads
on the given chiplet regards of which threads are selected.

Signed-off-by: Alistair Popple <alistair@popple.id.au>