Makefile.am: Allow static linking
Previously static linking did not work as the linker was not able to work out
that the HW units were referenced so it would throw that code away resulting in
the probing code finding no valid targets. This is fixed by adding some linker
flags to ensure the relevant code is always added to the binary.

Signed-off-by: Alistair Popple <alistair@popple.id.au>