libpdbg/cfam.c: Remove fsi2pib engine reset
Prior to every pib read and write we would reset the fsi2pib engine. This
was primarily added to improve reliability on P8 BMC's which do not allow
exclusive access to the GPIO registers for bitbanging FSI.

Newer BMC's have a built in kernel driver to allow FSI access and this
reset causes the fsi2pib engine to loose existing settings such as the true
mask register. To avoid this only reset the PIB master interface once at
startup.

Signed-off-by: Alistair Popple <alistair@popple.id.au>