libpdbg: Validate hardware units at build time
Hardware unit structs such as struct htm contain a struct pdbg_target
which can be passed around to create an abstraction. Currently there is
a bit of extra runtime code used to guard against developers not
declaring the struct pdbg_target member as the first member. If we can
enforce this at build time then a some runtime can be removed.

This patch enforces that the struct pdbg_target member of hardware
units is always the first member at build time.

The only caveat is that compiling with -O0 will result in a false
negative - the build succeeding when it should not. We should be using
litterally any other optimisation level so that isn't really a problem.

Signed-off-by: Cyril Bur <cyrilbur@gmail.com>