libpdbg: Move struct dt_node into struct pdbg_target
Currently we have this terrible mess of dt_nodes pointing to targets
and targets pointing to dt_nodes. This has been the cause of bugs due
to folling NULL pointers from a target expecting a node and vice versa.

Consolidating these structs means that there must be a node for a
target.

Signed-off-by: Cyril Bur <cyrilbur@gmail.com>