pdbg/htm: Enforce powersave=off
Core HTM has a nasty habit of not working if the stars haven't aligned
with the exact polar axis of Mars. To complicate matters if Zeus got up
on the incorrect side of the bed Core HTM will likely reflect his mood.

Core HTM requires that no threads be in powersave. It isn't clear if
this restriction only applies to threads on the core being traced on or
if this is a global thing. In order to be as safe as possible, pdbg
will enforce that powersave be disabled globally.

Signed-off-by: Cyril Bur <cyrilbur@gmail.com>