libpdbg: Add private data field for each targets
Client applications may need to store per-target data. Rather than having
to create their own data structures to do this it would be convenient if it
could be stored this in the target itself.

This patch adds a private data pointer to the struct pdbg_target definition
and provides a pair of accesses to set/get the private data pointer value.

Signed-off-by: Alistair Popple <alistair@popple.id.au>