pdbg: Make target selection explicit
We have now moved entirely to lazy probing of targets. This means client
applications now need to explicitly probe each child target they need
rather than explicitly marking everything they need prior to initialising
the library.

This leads to problems with device selection in pdbg as previously it would
run a command on every target which was found to exist after probing which
would ignore disabled targets. However disabling a target now prevents it
being probed even if child targets require it so pdbg can no longer set a
target to disabled just because it doesn't want to run a command on that
target.

Instead it needs to explicitly record which targets a command should be run
on. This is performed by marking each target as selected or not prior to
running commands.

Signed-off-by: Alistair Popple <alistair@popple.id.au>