src/htm.c: Only run on selected targets
The HTM does not call the same target iterators as the rest of the pdbg
application. Therefore the changes to device selection were not carried
over. Longer term we should change the HTM code to call the same iterators
as the rest of the pdbg application, but for the moment just copy the
device selection code across as the longer term fix really requires
rewritting the iterators to not use callbacks.

Signed-off-by: Alistair Popple <alistair@popple.id.au>