pdbg: add an expert mode
Add an --expert mode which contains dangerous options or things which
are "stateful" between pdbg invocations, where the release command and
possibly others need to be run to return the system to a normal state.

If you're not in expert mode, you just run a command and that's it.
More commands can be added to expert mode as we go.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>