p8: Add support for second processor on Firestone
The second processor module on Firestone is connected to a different HMFSI
port than Garrison. In theory we should/could probe every HMFSI port but
for simplicitly we really only need to probe the ones that we know to
actually be in use on various systems so just add the port used by
Firestone.

Signed-off-by: Alistair Popple <alistair@popple.id.au>