pdbg: Fix target selection when using I2C
I2C based access is unique in that to access the PIB on secondary processors we
need to traverse the PIB on the primary processor. This confuses our current
selection logic as every core becomes a child of the primary PIB. Hopefully our
selection logic will get rewritten soon, but in the meantime add a check to
ensure a cores immediate parent PIB matches the currently selected processor.

Signed-off-by: Alistair Popple <alistair@popple.id.au>