Remove "expert" mode
Based on feedback it seems "expert" mode was too vague and confusing for users,
so remove it. This means the user will be forced to manually manage the state of
the system which also seems to be too confusing for some users, but a future
patch will add user-friendly commands to the application which will change the
system state as required.

Signed-off-by: Alistair Popple <alistair@popple.id.au>