pdbg: Define common command/flag argument parsers
Parsers need to take a string argument and convert it into a specific type,
returning the value as a pointer to memory allocated by the parser (or NULL in
the case of an error). This just defines a couple of common parsers useful for
various commands.

It also defines a couple of convenience macros for easier usage of these
parsers.

Signed-off-by: Alistair Popple <alistair@popple.id.au>