libpdbg/chip.c: Fix logic error attempting to restore r0 and r1
When instruction ramming fails we attempt to restore r0 and r1. This is
unlikely to work as instruction ramming isn't likely to magically start
working if a previous attempt failed. Therefore it is only attempted once
before bailing. However a logic bug in the code meant it would try
indefinitely.

Signed-off-by: Alistair Popple <alistair@popple.id.au>