libpdbg/p8chip.c: ram state setup sequence match workbook
This makes a few changes to stop and ram procedure.

First of all, the existing thread_stop procedure is also setting up some
of the ram state. Change that to just do the stop sequence from the
workbook, and move the ram stuff into ram setup and destroy.

The workbook calls for inactive threads being rammed to set a thread
active state before ram mode is exited, in order for GPRs modified by
ramming to avoid getting lost. Currently the code does that in the
stop sequence before ram mode is activated.

The code also currently deasserts the thread active bit after exiting
ram mode, which is not part of the workbook, so this is no longer done.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>