htm: Fixed '-l' option to work as intended with new probing
This option was initially added for htm core tracing where we want the
core that the specified thread belongs to to be selected. The behaviour
was changed to only select the thread when the new probing was added.
Let's change the behaviour back, and select the pib, core and thread.

Signed-off-by: Rashmica Gupta <rashmica.g@gmail.com>
Acked-by: Michael Neuling <mikey@neuling.org>