libpdbg: Add default device-tree selection to libpdbg
For portability reasons libpdbg intentionally did not deal with the
selection and retrieval of the device tree for a given system. This
means that by design each application was responsible for selecting
and loading the correct device tree for the system it was on.

This is convenient for embedded systems as it removes the dependency
on some libc functions such as open. However there are currently no
embedded users of libpdbg and forcing end user applications to select
the correct device tree is likely to be error prone and inconsistent.

Instead this patch adds the ability for applications to call
pdbg_targets_init(NULL) and have libpdbg attempt to detect and load
the correct device tree for the current system. This doesn't change
any existing behaviour as applications may still pass an explicit
device-tree to pdbg_targets_init() which will always be used.

An environment variable is provided so users may override the
auto-detected device-tree when the application does not provide one.

This patch will bloat the overall size of the library as it will now
include all the device-trees. It also likely makes things harder for
embedded users. However if either of these is an issue we should be
able to easily add a build time option to exclude this and/or only
build in specific device-trees.

Signed-off-by: Alistair Popple <alistair@popple.id.au>
Reviewed-by: Amitay Isaacs <amitay@ozlabs.org>