pdbg: Remove default_backend() selection
Now that libpdbg can automatically select the correct device-tree the
pdbg application doesn't need code to do the same thing. By default
pdbg can just rely on the device-tree selected by libpdbg.

Unfortunately to maintain backwards compatibility for users explicitly
selecting a device-tree and device-tree node we need to retain most of
the code and continue linking the device-trees into the
application. It would be good to remove these one day so print a
deprecation warning when a user does explicitly select a backend.

Signed-off-by: Alistair Popple <alistair@popple.id.au>
Reviewed-by: Amitay Isaacs <amitay@ozlabs.org>