P8 ADU: Fix access sizes
Commit 302f249d95f0 ("adu: Add arugments for block size") introduced
the ability to specify an access size for the POWER8 and POWER9 ADU
functions for cache-inhibited access.

On POWER8 the ADU control register TSIZE encoding is slightly
different than for POWER9. Unfortunately the code for P8 was copied
from the P9 version but the encoding was not updated. This resulted in
ADU putmem commands writing twice as much memory as they were supposed
to on POWER8.

Signed-off-by: Alistair Popple <alistair@popple.id.au>
Reviewed-by: Amitay Isaacs <amitay@ozlabs.org>