libpdbg: Add virtual node target
This is used to construct system device tree representation.

Currently there are multiple device trees for a system (e.g. P9) based
on the backend where it's used.  This means one cannot refer to a hwunit
by specific path which is independent of backend.

System device tree is a system representation that allows to identify
hwunits uniquely by device tree path.  However, the backend device
trees are still separate.  To avoid keeping two different device trees
(system device tree and backend device tree) in memory, a system device
tree view is created on top of the backend device tree.  This system
device tree view is created using virtual nodes.  A virtual nodes
denotes the attachment point for a hw-unit in a system device tree view.
With multiple virtual nodes a system device tree view is overlaid on
top of the backend device tree.

Signed-off-by: Amitay Isaacs <amitay@ozlabs.org>
Reviewed-by: Alistair Popple <alistair@popple.id.au>