libpdbg: Create virtual nodes based on system-path property
To create a virtual node, "system-path" property is used to denote the
attachment point in the system device tree view.

It's also possible to create virtual nodes from backend device tree as
nodes without compatible property.

If a virtual node is mapped to a real node, then copy any properties
(except #address-cells and #size-cells) to the real node.

Signed-off-by: Amitay Isaacs <amitay@ozlabs.org>
Reviewed-by: Alistair Popple <alistair@popple.id.au>