libpdbg/host.c: Use reg property index of index to find chip-id
The host backend needs the hardware chip-id to find the correct
debugfs xscom file but incorrectly derives this from the index. This
means the chip path contains the actual chip-id rather than the index
which should be a logcal index starting at zero as it does for all
other backends.

This patch updates the host backend to use reg property for the
chip-id. A future patch will alter the device-trees to make this the
case.

Signed-off-by: Alistair Popple <alistair@popple.id.au>