pdbg: Make getmem always hexdump
The getmem command would not hexdump if printable characters were
found. This was useful for dumping text buffers such as the OPAL
console buffer. However when dumping other regions it is still common
for there to be printable characters which results in the output not
getting hexdumped.

Instead always hexdump, raw output can always be obtained using the
`--raw` option.

Signed-off-by: Alistair Popple <alistair@popple.id.au>