libpdbg.c: Fix address translation
If a target provides a translation callback it's expected that the
callback will return the absolute bus address to be used for the
access. However the caller of this callback was continuing to apply
translations to this absoulte address resulting in incorrect
translations in some cases. Instead short-circuit the translation
process.

Signed-off-by: Alistair Popple <alistair@popple.id.au>