libpdbg: Make pdbg_target_path() return a const string
Currently pdbg_target_path() allocates memory to return the path
into. As the target path is a convient indentifier when printing
information to user about a specific target it leads either to memory
leaks when called directly as part of a printf statement. The
alternative is lots of repeated code like the below, noting that the
assert can never be hit as there is already one in dt_get_path():

path = pdbg_target_path(tgt);
assert(path);
printf("%s", path);
free(path);

Solve this by adding a cache of the path to struct pdbg_target and return
that instead.

Signed-off-by: Alistair Popple <alistair@popple.id.au>