libpdbg: Run hwp only if target is enabled and hwp is defined
For bridge devices (fsi, pib, opb, etc.), the enabled check is on the
bridge devices and not on the children of the bridge devices.

For example, a core device can do a pib_read() during probe.  That means
core device is not yet enabled, but the parent pib is enabled.

Signed-off-by: Amitay Isaacs <amitay@ozlabs.org>