build: update instructions for hard-float BMCs
The previous patch was papering over the issue, badly. The cause of
the problem is that some OpenBMC builds have started to use hard float.
I didn't look hard enough and my ignorance of the ARM environment (and
the fact it "worked") made me think this was a more benign configuration
issue.

So update the build instructions to select the hard-float toolchain
when necessary. Confirmed this works on Debian.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Reviewed-by: Joel Stanley <joel@jms.id.au>