libpdbg/p10chip: include maintenance mode but not lsu quiesced in quiesced thread status
Maintenance mode is required for direct controls operations, so include
that in thread quiesced state. LSU quiesced is only required for RAM
mode where all threads in the core must be stopped, so remove that bit
from the status.

Change the structure of the code to be a little more self-documenting.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Reviewed-by: Amitay Isaacs <amitay@ozlabs.org>