libpdbg: fix p9chip and p10chip thread_stop
Have p9 and p10 thread_stop return 1 on failure, and have them set
thread->status to the status used when checking for quiesce, rather
than sampling it again on the way out.

This seems to improve some intermittent condition on POWER9 where
a thread was stopped successfuly but the subsequent status read
finds it no longer quiesced.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Joel Stanley <joel@jms.id.au>