gdbserver: reduce PR_INFO noise
Printing the first byte of the received packet in hex on its own line is
not very useful and clutters debug output. We print the full received
packet anyway which is generally quite readable.

Remove the first byte info line, and replace it with a debug line that
contains the entire hex string as hex bytes.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Reviewed-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>