gdbserver: lexer fixes
The lexer is coded in a funny way that ignores all packets that don't
match a command or interrupt to start with, and then expects a command
or ack/nack following an interrupt. This results in leading ack/nack not
being matched as such which is fine because they're ignored but makes
things confusing. But it also results in command errors for the command
following an interrupt.

All this complexity and problems seems to have arisen because the
trailing NUL termination is passed in to the buffer which has made
everything more complicated.

Fix this by not passing in the trailing NUL, and rewriting the lexer
so it always matches one of a command, interrupt, ack, or nack at
the high level. Command / crc reset is done at entry to the command
state.

Some additional checks and debugging messages are added.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Reviewed-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>