libpdbg: Remove enable_attn target command
enable_attn is only used by gdbserver and it can be implemented with
existing target register access. Implementing it for P9, P10, and
sbefifo backends results in duplication and P9/P10 cases in the sbefifo
backend.

Arguably it should be a middle-end function in the p*chip.c files which
is backend-agnostic. For now move it to gdbserver and we can look at
cleaning it up later.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Reviewed-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>