gdbserver: use read-modify-write for put_mem that is not 8-byte aligned
Not all targets have memory access that can support arbitrary byte
writes. Use RMW for put_mem commands that are not 8-byte aligned.

These helpers should possibly be moved into core code and this should
really either be done in the target accessor, or at least an alignment
capability should be exposed to the caller. But for now this will allow
sbefifo mem ops to work to set breakpoints (which requires 4-byte
writes).

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Reviewed-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>