gdbserver: check for attn using the SPATTN register
When POWER9/10 execute an attn instruction, they set a bit in the SPATTN
register and quiesce the thread.

This bit can be checked to confirm the thread hit an attn instruction,
rather than assuming a thread was quiesced because of attn. This makes
the gdb breakpoint code more robust in the presence of other direct
controls (e.g., host-based direct controls from OPAL).

This change also clears the SPATTN bit which clears the exception
condition that raises the IPOLL global interrupt. That interrupt seems
to be involved with the IPOLL interrupt storm and lock-up on POWER9
(although this does not fix it).

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Reviewed-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>