gdbserver: initial thread support
Add a command-specific pointer to the struct thread that gdbserver
can use, so it doesn't have to build its own structure of threads.

Initially keep the PIR value in here, as a demonstration.

Later PIR will be used to identify threads to the gdb client, and
more data will be added to the gdb_thread structure to support
debugging multiple threads.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Reviewed-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>