gdbserver: track attn enablement by breakpoints
attn-enabled is a per-core property that host software (e.g.,
skiboot) may change. gdbserver should not disable attn if it
was found to be enabled.

There are still unavoidable races where the host could change
between gdbserver wanting it enabled and disabled.

This also moves set_attn to iterate all targets in preparation
for multi-threaded debugging.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Reviewed-by: Joel Stanley <joel@jms.id.au>
Signed-off-by: Joel Stanley <joel@jms.id.au>