gdbserver: Add a note about the attn_handler service on the BMC
gdbserver's use of attn for breakpoints is not well integrated into the
rest of the firmware stack at the moment. attn firing can cause the BMC
to perform error handling (e.g., PEL, reboot).

Add a note to the README to disable the attn_handler.service if
breakpoints are to be used with gdbserver. While we're here update
the documentation to reflect that we no longer have to stop all
threads before running gdbserver.

Signed-off-by: Nicholas Piggin <npiggin@gmail.com>
Signed-off-by: Joel Stanley <joel@jms.id.au>