Revert class names of pib and fsi targets for odyssey chips
At present class name of odyssey pib and fsi targets are set
by appending "ody" to "pib" and "fsi" names to differntiate from
ddr4 pib, fsi targets to ddr5 pib and fsi targets.

It was done to limit the search for the pib/fsi target of the odyssey
chip when looping through all the pib and fsi target for the matching
proc and ocmb target. Not much advantage though but done for clarity.

The change in class names for ody pib and fsi targets are failing in
some of the sbe methods like sbe_set_state which are expecting
old pib and fsi class names.

Signed-off-by: Marri Devender Rao <devenrao@in.ibm.com>
Change-Id: Ibe589b73e8561c33d2166af812cd4f25aa59078e