Look for Rainier or Everest substring in Model
The expectation was that the /proc/device-tree/model file would precisely
indicate the system model, such as Rainier 2u, Rainier 4U, or Everest.
Previously, we relied on exact matching of the entire string pattern to load the
corresponding Odyssey dtb files. However, in some test systems, additional
characters caused the existing conditions to fail, resulting in
misidentification as a DDR4 system. To address this issue, the code was modified
to search for substrings instead of exact matches. This enabled the code to
function as intended.

The ecmdquery chips were tested, successfully probing and loading the necessary
Odyssey backend targets