Fix corner case
In contains_substring_ignoring_case method, if the string that we are
looking for is longer than the source string, return with string not
found.  Earlier the char array index went out of bounds due to invalid
arthimetic operation on the index.

Tested and found the method working as expected.

Signed-off-by: deepakala-k <deepakala.karthikeyan@ibm.com>